move neopointworldtext to new file
