+++ src/game/client/CMakeLists.txt
@@ -1571,6 +1571,7 @@ target_sources_grouped(
 set(UNITY_SOURCE_NEO_SRC_CLIENT
     neo/c_neo_message.cpp
     neo/c_neo_npc_dummy.cpp
+    neo/c_neo_point_world_text.cpp
     neo/c_neo_player.cpp
     neo/c_neo_te_tocflash.cpp
     neo/neo_fixup_glshaders.cpp
@@ -1589,6 +1590,7 @@ target_sources_grouped(
     FILES
     neo/c_neo_message.h
     neo/c_neo_npc_dummy.h
+    neo/c_neo_point_world_text.h
     neo/c_neo_player.h
     neo/c_neo_te_tocflash.h
     neo/neo_fixup_glshaders.h
